+++ changelog.md
@@ -1,3 +1,2 @@
-- Added disable voicechat button
-- Added disconnect indicator
-- Fixed sound coming from the wrong side when in F5
+- Added mono sound config option
+- Added option to hide icons